refactor: move DashboardContent to commonMain
Splits DashboardContent out of DashboardScreen.kt into commonMain so it
can be exercised by PreviewScreenshotGeneratorTest on the JVM target.
