+++ feature/dashboard/build.gradle.kts
@@ -4,12 +4,14 @@ plugins {
 
 kotlin {
     sourceSets {
+        commonMain.dependencies {
+            implementation(project(":core:ui"))
+            implementation(project(":shared"))
+        }
         wasmJsMain.dependencies {
             implementation(project(":core:auth"))
             implementation(project(":core:common"))
             implementation(project(":core:network"))
-            implementation(project(":core:ui"))
-            implementation(project(":shared"))
         }
     }
 }
